# Add array.array to array, bytearray to bytes and write_int improvement