--- a/fastavro/_validation.pyx
+++ b/fastavro/_validation.pyx
@@ -1,5 +1,6 @@
 # cython: language_level=3str
 
+import array
 import numbers
 try:
     from collections.abc import Mapping, Sequence
@@ -84,7 +85,7 @@ cdef inline bint validate_array(
     str parent_ns='',
     bint raise_errors=True,
 ) except -1:
-    if not isinstance(datum, Sequence) or is_str(datum):
+    if not isinstance(datum, (Sequence, array.array)) or is_str(datum):
         return False
 
     for d in datum: