--- a/fastavro/_validation_py.py
+++ b/fastavro/_validation_py.py
@@ -1,3 +1,4 @@
+import array
 import numbers
 try:
     from collections.abc import Mapping, Sequence
@@ -196,7 +197,7 @@ def validate_array(
         If true, raises ValidationError on invalid data
     """
     return (
-            isinstance(datum, Sequence) and
+            isinstance(datum, (Sequence, array.array)) and
             not is_str(datum) and
             all(_validate(datum=d, schema=schema['items'],
                           named_schemas=named_schemas,