--- a/tests/test_complex.py
+++ b/tests/test_complex.py
@@ -134,7 +134,7 @@ def test_array_from_tuple():
 def test_array_from_array():
     schema = {"type": "array", "items": "int"}
     using_list = serialize(schema, [1, -2, 3])
-    using_array = serialize(schema, array.array("q", [1, -2, 3]))
+    using_array = serialize(schema, array.array("l", [1, -2, 3]))
     assert using_list == using_array
     assert deserialize(schema, using_list) == [1, -2, 3]