--- a/tests/test_types.py
+++ b/tests/test_types.py
@@ -94,22 +94,25 @@ def test_int_in_string_raises():
         serialize(schema, *records)
 
 
-def test_int_bytes():
-    test_cases = [(0, b"\x00"),
-                  (-1, b"\x01"), (1, b"\x02"),
-                  (-2, b"\x03"), (2, b"\x04"),
-                  (2147483647, b"\xfe\xff\xff\xff\x0f"),
-                  (-2147483648, b"\xff\xff\xff\xff\x0f"),
-                  (9223372036854775807,
-                   b"\xfe\xff\xff\xff\xff\xff\xff\xff\xff\x01"),
-                  (-9223372036854775808,
-                   b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\x01")
-                  ]
-
+@pytest.mark.parametrize(  #
+    ("value", "binary"),  #
+    [
+        (0, b"\x00"),
+        (-1, b"\x01"),
+        (1, b"\x02"),
+        (-2, b"\x03"),
+        (2, b"\x04"),
+        (2147483647, b"\xfe\xff\xff\xff\x0f"),
+        (-2147483648, b"\xff\xff\xff\xff\x0f"),
+        (9223372036854775807, b"\xfe\xff\xff\xff\xff\xff\xff\xff\xff\x01"),
+        (-9223372036854775808, b"\xff\xff\xff\xff\xff\xff\xff\xff\xff\x01"),
+    ])
+def test_int_binary(value, binary):
     schema = {"type": "long"}
+    buffer = BytesIO()
 
-    for value, encoded in test_cases:
-        buffer = BytesIO()
-        fastavro.schemaless_writer(buffer, schema, value)
-        serialized = buffer.getvalue()
-        assert serialized == encoded, "Invalid integer encoding encountered."
+    fastavro.schemaless_writer(buffer, schema, value)
+    assert buffer.getvalue() == binary, "Invalid integer encoding."
+
+    deserialized = fastavro.schemaless_reader(BytesIO(binary), schema)
+    assert deserialized == value, "Invalid integer decoding."