First of all, sorry for adding too many items in one PR.  They're well separated in commits, so I can easily drop one or more parts.
As a second remark, I should probably add I have no experience with cython, so I'll be happy to read your comments.

The first change allows for python's `array.array` to be serialized to an array.
The second change allows for a `bytearray` to be serialized to bytes.  I had to move some logic to `write_utf8` to avoid performance loss.
The last commit introduces a small change in `write_int`, that improves cython performance for me.

I also added a `timeit`-based benchmark, to easily follow up the effects:

original:
```
Array from list                   100 records, best of   1000
========================================================
write fastavro               100 records, best of   1000:  14.400871 ms
read  fastavro               100 records, best of   1000:  15.655241 ms
write schemaless fastavro    100 records, best of   1000:  14.669046 ms
read  schemaless fastavro    100 records, best of   1000:  13.781189 ms
validate fastavro            100 records, best of   1000:  15.007082 ms

real    1m16.688s
user    1m15.806s
sys     0m0.705s
```

after array.array and bytearray addition:
```
Array from list                   100 records, best of   1000
========================================================
write fastavro               100 records, best of   1000:  14.400178 ms
read  fastavro               100 records, best of   1000:  15.634401 ms
write schemaless fastavro    100 records, best of   1000:  14.759453 ms
read  schemaless fastavro    100 records, best of   1000:  13.944073 ms
validate fastavro            100 records, best of   1000:  15.596337 ms

real    1m20.245s
user    1m19.191s
sys     0m0.752s
```

after write_int change:
```
Array from list              100 records, best of   1000
========================================================
write fastavro               100 records, best of   1000:  11.314640 ms
read  fastavro               100 records, best of   1000:  15.760850 ms
write schemaless fastavro    100 records, best of   1000:  11.621763 ms
read  schemaless fastavro    100 records, best of   1000:  13.935867 ms
validate fastavro            100 records, best of   1000:  15.673204 ms

real    1m11.119s
user    1m10.324s
sys     0m0.626s
```

array.array vs list:
```
Array from list              100 records, best of    100
========================================================
write fastavro               100 records, best of    100:  11.315835 ms
read  fastavro               100 records, best of    100:  15.406354 ms
write schemaless fastavro    100 records, best of    100:  11.669317 ms
read  schemaless fastavro    100 records, best of    100:  14.000947 ms
validate fastavro            100 records, best of    100:  15.810282 ms

Array from array             100 records, best of    100
========================================================
write fastavro               100 records, best of    100:   9.927711 ms
read  fastavro               100 records, best of    100:  15.353774 ms
write schemaless fastavro    100 records, best of    100:  10.219035 ms
read  schemaless fastavro    100 records, best of    100:  14.010455 ms
validate fastavro            100 records, best of    100:  17.267594 ms
```


Validation for python arrays is slower than for lists, and could be improved for specific cases based on `array.itemsize` and signedness and/or cython memoryviews.

Finally, `write_python_array` probably could be improved by also adding cases where for example `float` isn't the exact type, but part of a union.