@scottbelden 2020-08-17
Thanks! I'll probably get this out in a release today or tomorrow.